fix: 修复 vxe-table 类型定义无效的问题
