+++ index.ts
@@ -5,21 +5,6 @@ import type ExcelJS from 'exceljs'
 let globalVxetable: VXETableCore
 let globalExcelJS: any
 
-declare module 'vxe-table' {
-  export namespace VxeTableDefines {
-    export interface ExtortSheetMethodParams {
-      workbook: ExcelJS.Workbook;
-      worksheet: ExcelJS.Worksheet;
-    }
-    export interface ColumnInfo {
-      _row: any;
-      _colSpan: number;
-      _rowSpan: number;
-      childNodes: VxeTableDefines.ColumnInfo[];
-    }
-  }
-}
-
 const defaultHeaderBackgroundColor = 'f8f8f9'
 const defaultCellFontColor = '606266'
 const defaultCellBorderStyle = 'thin'
